Package relocation and build improvements
Simplified the package name to `com.cyberscout.auth0`. Similar rename of the
Maven group. Added the Maven plugin to support installing the library to the
local Maven repo. Removed the Gradle Spring Boot plugin, which is only needed
by Spring Boot applications, and was interfering with building the project.
Instead, used the Spring Boot BOM directly to get the Spring dependency
versions.