Some refactoring of properties and documentation
Renamed the `clientId` and `clientSecret` properties to simply `id` and
`secret`, which will read better (since they are already under the `client`
sub-property).

Removed the `client.enabled` property, since it can be derived from the presence
of the `client.id` and `client.secret` properties. The
`client.authenticationEnabled` property can now serve the same purpose, if a
single flag is needed.

Fully documented the `Auth0Properties` class.

Fixed the static factory method of the `Auth0ManagementContext` class. The
`apiId` parameter was not needed, since it's always `management`. This further
separates the static factory method from the one in the super class.